requirements-dev.txt
black==22.12.0
isort==5.12.0
ruff==0.14.6
mypy==1.5.0
pre-commit==3.2.0
pytest-cov==4.1.0